refactor(cmd_cleanup): promote to directory, split tests (#1266)
- src/cmd_cleanup.rs: 776 → src/cmd_cleanup/mod.rs (479 LOC)
- New src/cmd_cleanup/tests.rs (297 LOC)

mod.rs still exceeds 400 LOC; needs follow-up source split.

Pure refactor: 14 tests pass.
